+++ src/components/content/medications/patient-medications-tabbed.tsx
@@ -4,6 +4,7 @@ import type {
 } from "@/components/core/filter-bar/filter-bar-types";
 import cx from "classnames";
 import { useEffect, useState } from "react";
+import { MedsHistoryTempProps } from "@/components/content/medications-table-base";
 import {
   BadgeOtherProviderMedCount,
   OtherProviderMedsTable,
@@ -23,29 +24,41 @@ import { uniq } from "@/utils/nodash/fp";
 export type PatientMedicationsTabbedProps = {
   className?: string;
   forceHorizontalTabs?: boolean;
-} & SubsetOtherProviderTableProps;
-type SubsetOtherProviderTableProps = Pick<
+} & TabbedContentProps;
+
+type TabbedContentProps = Pick<
   OtherProviderMedsTableProps,
   "hideAddToRecord" | "handleAddToRecord"
->;
+> &
+  MedsHistoryTempProps;
 
 // We use getPanelClassName on all tabs except for the other-provider-records
 // tab because without the FilterBar there is no margin between tab and panel
 // when md - lg sized.
 const tabbedContent = (
-  otherProviderTableProps: SubsetOtherProviderTableProps
+  tabbedContentProps: TabbedContentProps
 ): TabGroupItem<MedicationStatementModel>[] => [
   {
     key: "medication-list",
     getPanelClassName: (sm: boolean) => (sm ? "ctw-mt-0" : "ctw-mt-2"),
     display: () => "medication list",
-    render: () => <ProviderMedsTable />,
+    render: () => (
+      <ProviderMedsTable
+        onOpenHistoryDrawer={tabbedContentProps.onOpenHistoryDrawer}
+        onAfterOpenHistoryDrawer={tabbedContentProps.onAfterOpenHistoryDrawer}
+      />
+    ),
   },
   {
     key: "inactive-provider-records",
     getPanelClassName: (sm: boolean) => (sm ? "ctw-mt-0" : "ctw-mt-2"),
     display: () => "inactive",
-    render: () => <ProviderInactiveMedicationsTable />,
+    render: () => (
+      <ProviderInactiveMedicationsTable
+        onOpenHistoryDrawer={tabbedContentProps.onOpenHistoryDrawer}
+        onAfterOpenHistoryDrawer={tabbedContentProps.onAfterOpenHistoryDrawer}
+      />
+    ),
   },
   {
     key: "other-provider-records",
@@ -55,13 +68,15 @@ const tabbedContent = (
         <BadgeOtherProviderMedCount />
       </>
     ),
-    render: () => <OtherProviderMedsTableTab {...otherProviderTableProps} />,
+    render: () => <OtherProviderMedsTableTab {...tabbedContentProps} />,
   },
 ];
 
 export function OtherProviderMedsTableTab({
   handleAddToRecord,
   hideAddToRecord,
+  onOpenHistoryDrawer,
+  onAfterOpenHistoryDrawer,
 }: PatientMedicationsTabbedProps) {
   const [filters, setFilters] = useState<FilterChangeEvent>({});
   const [records, setRecords] = useState<MedicationStatementModel[]>([]);
@@ -120,6 +135,8 @@ export function OtherProviderMedsTableTab({
         hideAddToRecord={hideAddToRecord}
         showDismissed={showDismissed}
         showInactive={showInactive}
+        onOpenHistoryDrawer={onOpenHistoryDrawer}
+        onAfterOpenHistoryDrawer={onAfterOpenHistoryDrawer}
       />
     </>
   );
@@ -137,13 +154,16 @@ export function OtherProviderMedsTableTab({
 export function PatientMedicationsTabbed({
   className,
   forceHorizontalTabs = false,
-  ...otherProviderTableProps
+  ...tabbedContentProps
 }: PatientMedicationsTabbedProps) {
-  const tabItems = tabbedContent(otherProviderTableProps);
+  const tabItems = tabbedContent(tabbedContentProps);
 
   return (
     <CTWBox.StackedWrapper
-      className={cx("ctw-patient-medications ctw-space-y-3", className)}
+      className={cx(
+        "ctw-patient-resource-component ctw-patient-medications ctw-space-y-3",
+        className
+      )}
     >
       <TabGroup
         content={tabItems}